% Generated by build script
\name{giss_global_temp}
\alias{giss_global_temp}
\title{Download and parse GISTEMP global temperature anomalies}
\usage{
giss_global_temp()
}
\arguments{
\item{dataset}{One of "global" (land+ocean), "land", "southern", "northern".}
}
\value{tibble with year and monthly/seasonal temperature anomaly columns}
\description{Download and parse GISTEMP global temperature anomalies}
