% Generated by build script
\name{giss_trend}
\alias{giss_trend}
\title{Get GISTEMP trend summary (first and last decade averages)}
\usage{
giss_trend()
}
\arguments{
\item{dataset}{One of "global", "land", "northern", "southern". Default "global".}
}
\value{tibble: period, avg_annual_anomaly}
\description{Get GISTEMP trend summary (first and last decade averages)}
