% Generated by build script
\name{ntrs_recent}
\alias{ntrs_recent}
\title{Get recent NASA Technical Reports}
\usage{
ntrs_recent()
}
\arguments{
\item{page_size}{Number of results (default 25)}
\item{type}{STI type filter: "TECHNICAL_REPORT", "CONFERENCE_PAPER", "REPRINT", etc.}
}
\value{tibble: id, title, sti_type, distribution, created, authors}
\description{Get recent NASA Technical Reports}
